# openapi.json
{"openapi":"3.0.0","paths":{"/health":{"get":{"operationId":"AppController_health","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]},"patch":{"operationId":"AuthController_updateMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/auth/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/orders":{"post":{"operationId":"OrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["Orders"]},"get":{"operationId":"OrdersController_list","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/orders/summary":{"get":{"operationId":"OrdersController_summary","parameters":[],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/orders/page":{"get":{"operationId":"OrdersController_page","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"status","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"limit","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"from","required":false,"in":"query","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/orders/{id}":{"get":{"operationId":"OrdersController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Orders"]}},"/orders/{id}/complete":{"post":{"operationId":"OrdersController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["Orders"]}},"/orders/{id}/attach-customer":{"post":{"operationId":"OrdersController_attach","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AttachCustomerDto"}}}},"responses":{"201":{"description":""}},"tags":["Orders"]}},"/orders/{id}/return":{"post":{"operationId":"OrdersController_returnOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["Orders"]}},"/orders/{id}/reprint":{"post":{"operationId":"OrdersController_reprint","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Orders"]}},"/orders/{id}/cancel":{"post":{"operationId":"OrdersController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CancelOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["Orders"]}},"/print-jobs/next":{"get":{"operationId":"PrintingController_next","parameters":[],"responses":{"200":{"description":""}},"tags":["Printing"]}},"/print-jobs/{id}/success":{"post":{"operationId":"PrintingController_success","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrintJobClaimDto"}}}},"responses":{"201":{"description":""}},"tags":["Printing"]}},"/print-jobs/{id}/failure":{"post":{"operationId":"PrintingController_failure","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrintJobFailureDto"}}}},"responses":{"201":{"description":""}},"tags":["Printing"]}},"/notifications/zalo/status":{"get":{"operationId":"ZaloController_status","parameters":[],"responses":{"200":{"description":""}},"tags":["Zalo"]}},"/notifications/zalo/connect":{"post":{"operationId":"ZaloController_connect","parameters":[],"responses":{"201":{"description":""}},"tags":["Zalo"]}},"/notifications/zalo/disconnect":{"post":{"operationId":"ZaloController_disconnect","parameters":[],"responses":{"201":{"description":""}},"tags":["Zalo"]}},"/notifications/zalo/test":{"post":{"operationId":"ZaloController_test","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestZaloDto"}}}},"responses":{"201":{"description":""}},"tags":["Zalo"]}},"/settings/loyalty":{"get":{"operationId":"SettingsController_loyalty","parameters":[],"responses":{"200":{"description":""}},"tags":["Settings"]},"put":{"operationId":"SettingsController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateLoyaltySettingDto"}}}},"responses":{"200":{"description":""}},"tags":["Settings"]}},"/services":{"get":{"operationId":"ServicesController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["Services"]},"post":{"operationId":"ServicesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServiceDto"}}}},"responses":{"201":{"description":""}},"tags":["Services"]}},"/services/all":{"get":{"operationId":"ServicesController_all","parameters":[],"responses":{"200":{"description":""}},"tags":["Services"]}},"/services/{id}":{"patch":{"operationId":"ServicesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServiceDto"}}}},"responses":{"200":{"description":""}},"tags":["Services"]}},"/customers/search":{"get":{"operationId":"CustomersController_search","parameters":[],"responses":{"200":{"description":""}},"tags":["Customers"]}},"/customers":{"get":{"operationId":"CustomersController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["Customers"]}},"/customers/{id}":{"get":{"operationId":"CustomersController_get","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Customers"]},"patch":{"operationId":"CustomersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCustomerDto"}}}},"responses":{"200":{"description":""}},"tags":["Customers"]}},"/customers/{id}/loyalty-adjust":{"post":{"operationId":"CustomersController_loyaltyAdjust","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoyaltyAdjustDto"}}}},"responses":{"201":{"description":""}},"tags":["Customers"]}},"/expenses":{"post":{"operationId":"ExpensesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExpenseDto"}}}},"responses":{"201":{"description":""}},"tags":["Expenses"]},"get":{"operationId":"ExpensesController_list","parameters":[{"name":"page","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"limit","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"from","required":false,"in":"query","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","schema":{"type":"string"}},{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"paymentMethod","required":false,"in":"query","schema":{"type":"string"}},{"name":"includeVoided","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}}],"responses":{"200":{"description":""}},"tags":["Expenses"]}},"/expenses/{id}":{"patch":{"operationId":"ExpensesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateExpenseDto"}}}},"responses":{"200":{"description":""}},"tags":["Expenses"]}},"/expenses/{id}/void":{"post":{"operationId":"ExpensesController_void","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoidExpenseDto"}}}},"responses":{"201":{"description":""}},"tags":["Expenses"]}},"/shifts":{"get":{"operationId":"ShiftsController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["Shifts"]}},"/shifts/current":{"get":{"operationId":"ShiftsController_current","parameters":[],"responses":{"200":{"description":""}},"tags":["Shifts"]}},"/shifts/open":{"post":{"operationId":"ShiftsController_open","parameters":[],"responses":{"201":{"description":""}},"tags":["Shifts"]}},"/shifts/{id}/close":{"post":{"operationId":"ShiftsController_close","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloseShiftDto"}}}},"responses":{"201":{"description":""}},"tags":["Shifts"]}},"/dashboard/today":{"get":{"operationId":"DashboardController_today","parameters":[],"responses":{"200":{"description":""}},"tags":["Dashboard"]}},"/dashboard/range":{"get":{"operationId":"DashboardController_range","parameters":[],"responses":{"200":{"description":""}},"tags":["Dashboard"]}},"/audit-logs":{"get":{"operationId":"AuditController_list","parameters":[{"name":"page","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"limit","required":false,"in":"query","schema":{"$ref":"#/components/schemas/Object"}},{"name":"action","required":false,"in":"query","schema":{"type":"string"}},{"name":"entityType","required":false,"in":"query","schema":{"type":"string"}},{"name":"userId","required":false,"in":"query","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Audit"]}},"/users":{"get":{"operationId":"UsersController_list","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]},"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/{id}":{"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/{id}/reset-password":{"post":{"operationId":"UsersController_reset","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}}},"info":{"title":"ZUZU Laundry API","description":"Internal API contract","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"zuzu_session"}},"schemas":{"LoginDto":{"type":"object","properties":{"usernameOrPhone":{"type":"string"},"password":{"type":"string"},"remember":{"type":"boolean"}},"required":["usernameOrPhone","password"]},"UpdateProfileDto":{"type":"object","properties":{"name":{"type":"string"},"phone":{"type":"object"}},"required":["name"]},"ChangePasswordDto":{"type":"object","properties":{"currentPassword":{"type":"string"},"newPassword":{"type":"string"}},"required":["currentPassword","newPassword"]},"CreateOrderDto":{"type":"object","properties":{"phone":{"type":"string"},"customerName":{"type":"string"},"customerAddress":{"type":"string"},"note":{"type":"string"},"dueDate":{"type":"string"},"duePeriod":{"type":"string"},"deliveryAddress":{"type":"string"},"serviceIds":{"type":"array","items":{"type":"string"}},"customerId":{"type":"string"},"customerUnknown":{"type":"boolean"}},"required":["customerUnknown"]},"Object":{"type":"object","properties":{}},"CompleteOrderDto":{"type":"object","properties":{"items":{"type":"array","items":{"type":"string"}},"discount":{"type":"number"},"expectedUpdatedAt":{"type":"string"}},"required":["items"]},"AttachCustomerDto":{"type":"object","properties":{"phone":{"type":"string"},"name":{"type":"string"}},"required":["phone"]},"ReturnOrderDto":{"type":"object","properties":{}},"CancelOrderDto":{"type":"object","properties":{}},"PrintJobClaimDto":{"type":"object","properties":{"claimToken":{"type":"string"}},"required":["claimToken"]},"PrintJobFailureDto":{"type":"object","properties":{"claimToken":{"type":"string"},"error":{"type":"string"}},"required":["claimToken","error"]},"TestZaloDto":{"type":"object","properties":{"phone":{"type":"string"}},"required":["phone"]},"UpdateLoyaltySettingDto":{"type":"object","properties":{}},"CreateServiceDto":{"type":"object","properties":{}},"UpdateServiceDto":{"type":"object","properties":{}},"UpdateCustomerDto":{"type":"object","properties":{}},"LoyaltyAdjustDto":{"type":"object","properties":{}},"CreateExpenseDto":{"type":"object","properties":{"amount":{"type":"number"},"category":{"type":"string"},"description":{"type":"string"},"expenseDate":{"type":"string"},"paymentMethod":{"type":"string"},"receiptUrl":{"type":"string"}},"required":["amount","category","description","expenseDate","paymentMethod"]},"UpdateExpenseDto":{"type":"object","properties":{"amount":{"type":"number"},"category":{"type":"string"},"description":{"type":"string"},"expenseDate":{"type":"string"},"paymentMethod":{"type":"string"},"receiptUrl":{"type":"string"}}},"VoidExpenseDto":{"type":"object","properties":{}},"CloseShiftDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"ResetPasswordDto":{"type":"object","properties":{}}}}}